Show file name in unit test output
When adding a new .xaml file I always forget to update CommonFiles.swr to include it. Happily, we have a unit test that catches this problem. However, it doesn't currently tell you which file you need to update to correct the problem--and that's another thing I never remember.

This commit updates the test to be explicit about needing to update CommonFiles.swr.