Add LaunchTargetPropertyPageValueProvider
Fundamentally, the `LaunchTargetPropertyPageValueProvider` maps from a debugger launch target (also known as the "command name") to the name of a property page that defines the properties relevant to that target, and back. Given a launch target, it will check the debugger property pages for one with matching "CommandName" metadata--this occurs in the "get property" operation. Given a page name, it will look up the corresponding page, extract the "CommandName" metadata, and update the active launch profile to use that launch target. This occurs in the "set property" operation.

The idea is that there will be a debugger "parent" page that manages the active profile and that profile's launch target. When the launch target changes, we need to identify and show the corresponding "child" page with the properties relevant to that launch target.