Add "executable" and "project" debug pages
Define debug property pages for the "executable" and "project" launch targets. Each launch target has its own `Rule` specified in a .xaml file, with the debugger command specified in the `Rule` metadata. Along with the support for "parent" and "child" debugger property pages added in https://devdiv.visualstudio.com/DevDiv/_git/CPS/pullrequest/248579 this allows the user to change between the two different launch targets and see the set of properties update appropriately.

Note that many of the relevant properties have not been added to the .xaml files as we don't yet have value providers for them on the back end. These will be added in a future commit.