Don't throw if unknown active framework is "Unsupported"
Fixes #6204
Fixes AB#1126588

We are seeing some customer NFEs around this check where the active target framework is "Unsupported,Version=v0.0". The validation is theoretically correct, but the consequences of continuing in such a scenario are minor-to-non-existant. Therefore I'm leaving it as an assert to assist during development of this code.

The real fix for this problem would be covered by #6183.