Enforce consistent properties between resolved/unresolved dependency pairs
For every dependency type, we have a resolved and unresolved rule.

This commit ensures that both rules in every pair have the same visible
properties, and that those properties have the same display name and description.

Long term we should split this so we have a single rule for the browse object
with all-visible properties, and other rules for internal processing with no
visible properties.

Adds a unit test to ensure consistency.