Update page templates
Fixes #6225.

CPS assumes each "debugger" Rule (with `PageTemplate="debugger"`) should be shown in the launch target drop down in the toolbar, and expects that each provides the GUID and ID of the related VS debugger command in their metadata. Rules which do not are associated with default GUIDs and IDs.

Our debugger rules don't provide the GUID and IDs because they provide a debug target (e.g., Executable, Project, IIS, IISExpress, etc.) instead. For the purposes of the debugger they aren't directly "startable" on their own--they need to be associated with a launch profile that provides all the relevant settings. So we don't want them to appear on this menu.

I can't simply update CPS to filter out "debugger" pages that lack the GUID and ID, because in theory other pages are usable by the debugger with the default values and we want to keep showing those.

Instead I've updated the property page code in CPS ([PR 255874](https://devdiv.visualstudio.com/DefaultCollection/DevDiv/_git/CPS/pullrequest/255874)) to look for a new kind of page (`PageTemplate="commandNameBasedDebugger"`), and this commit updates the appropriate .xaml files in our repo to use this new page template.