There's four main knobs that control the amount of generated code for rules:
XamlPropertyRule/XamlPropertyRuleNoCodeBehind item: Whether or not to generate any code whatsoever.
RuleInjection: Whether or not to generate strongly typed extension methods that that hang off ProjectProperties.
DataAccess: Whether to generate the rule, or just the schema/property names.
CreateFallbackRule: Whether to generate a rule that loads the rule from inside the assembly if the one on disk is missing.

Went through all rules and determined which ones we were using and which we were not, and toggled them based on that. I'm not touching CreateFallbackRule until I've confirmed that we're not actually relying on this in some environments where our rules are missing (1ES, for example).

This reduces our dll size by ~15% from just under 1 MB down to 837 KB.