Track dependency diagnostic level at snapshot level
The concept of 'diagnostic level' was introduced in #6288. Previously, warnings were only shown on dependencies when they were marked as unresolved. Now we can have multiple levels.

This commit tracks the maximum visible diagnostic level in a snapshot at the snapshot level (both targeted and cross-target), and passes that value to code that's responsible for placing an icon in the tree.

This fixes an issue where a child could have a warning icon visible, but the parent would not (seen in https://github.com/dotnet/project-system/issues/6275#issuecomment-653365600).