Replace ActiveConfiguredProject usage with new APIs
Fixes: https://devdiv.visualstudio.com/DevDiv/_workitems/edit/1042733

ActiveConfiguredProject.Value throws a MEF exception when accessing a service that is not applicable to the capabilities of the ConfiguredProject. Rather than push MEF exceptions onto all callers of this, this replaces ActiveConfiguredProject usage with two new APIs that work similar to IVsService<T>, where it returns null when the configuration does not provide the service. The majority of usages of this API ask for a service that always exists so do not need this API.

This also simplifies another pattern; pulling a collection of values from the active ConfiguredProject, previously this needed an inner class to wrap the values.

Usage:

public FooService(IActiveConfiguredValue<IService> service);

...

service.Value?.Method();

...

public FooService(IActiveConfiguredValues<IService> services);

....

foreach (Lazy<T> service in services)
{
   service.Value.Method();
}