Downgrade analyzers during dev builds to warnings
In previous versions of VS certain analyzers (like new API or missing file headers) would error but not block the build because only the IDE produced the error. Now editorconfig is plumbed through the build to the compiler, it now stops the compiler from producing outputs and hence blocks F5. This can be annoying.

This downgrades the majority of the analyzers to warnings. They continue to get prompted to errors during CI/official builds and hence will block merge until they are fixed.