Fix root graft tree provider GetPath logic
Root graft project tree providers should only return a value for GetPath for tree nodes that they own. Previously the assumption was that the method was only passed its own members, but this is not the case.

This fixes a problem where, when the Dependencies tree was not enabled, the "Copy Path" command on import tree items would result in the pseudo-path (">123") being copied rather than the actual file path.