Support "Environment Variables" property
Add basic support for the "Environment variables" property as used by various debugger property pages. This commit updates `ActiveLaunchProfileCommonValueProvider` to serialize the `ILaunchProfile.EnvironmentVariables` dictionary to a string when the value is retrieved, and deserialize it from a string back to a dictionary when it is set.

In serializing to a string we need some way to separate the environment variable name from its value, and also some way to separate one environment variable from the next. Here we use an equal sign `=` for the former, and a comma `,` for the latter. These were chosen as they were unlikely to appear in variable names or values and wouldn't need to be escaped. When we do need to escape them, we do so using a forward slash `/`.