Merge "GenericDependency" and "Dependency" tree flags
GenericDependency was apparently intended to identify dependencies provided by the project system, as opposed to those from third party extenders. The only usage of this was in the handling of `SchemaItemType`, to force the value to `Folder.PrimaryDataSourceItemType` for items that do not come from XAML rules, and therefore cannot have their browse objects populated with a rule schema name.

In practice, such extensions will pass `null` for `SchemaItemType`, which will default to the same value anyway.