Dependencies don't need VirtualFolder and BubbleUp flags
VirtualFolder is only needed on the root node (where it's already specified), target group nodes (where it's been added in this commit) and dependency type group nodes (where it's already specified).

BubbleUp only makes sense on the "Dependencies" node, to make it stick to the top of the project tree. It's already specified there.