Call GetProperty directly
When we're looking up a property editor by ID we currently get the property via the `TryGetPropertyAndIndex` extension method. We need to do things this way in `UIPropertyByIdProducer` because the index is part of the property data we need to return. However, here we're just getting the property in order to get to the editor, and we don't actually care about the index. Calling `GetProperty` is preferable because it can just look up the property by name in a dictionary, while `TryGetPropertyAndIndex` must do a linear traversal of all the properties in the rule.