Refactor logic to read supported frameworks
This refactors the logic to read supported frameworks, and removes the CPS implementation from parsing/understanding the "moniker" being based through from SupportedTargetFramworksEnumProvider.

This is a precursor to enabling us to bind to the "TargetFramework" property in https://github.com/dotnet/project-system/pull/6607.

Before this change;

1) The moniker returned from SupportedTargetFrameworksEnumProvider was being parsed and filtered. This breaks when that moniker is a "friendly identifier" such as "net5.0-windows".

2) Web/Legacy/CPS filtering was all intertwined. This separates the logic into separate types.

3) This fixes a slight bug where the display name for .NET Framework-only targets in SDK-based projects would show as ".NETFramework,v4.5" if the TFM wasn't installed.

There is not behavior change in all scenarios except for 3).