Delegate to new project properties editor via feature flag
When the user opts in to feature "CPS.UpdatedProjectPropertiesDesigner.Local" we
redirect requests for the legacy project editor to the new project editor.

There is currently a bug where the editor only appears on the second launch. For
now this commit adds value by enabling development of the new property pages
without requiring a Codespaces environment.