Enable property ProjectAssetFile for Roslyn.
Roslyn team needs to get the path of project.assets.json
for specific projects. This is part of the new feature
Remove Unused Reference.

This change pushes to Roslyn the msbuild evaluation ProjectAssetsFile
containing the absolute path for project.assets.json so they can get
this info on their side.