Allow the user to disable the feature symbolic renaming on File-rename
It seems that this option is misleading since we are enabling the
feature unconditionally. People might get confused when trying to
disable the prompt and still seeing changes in your code.

This change replaces the dialog box to allow the option "Don't Ask
Again" so users can disable the prompt message from here.
CheckUserConfirmation() now returns false and that means the feature is
disabled by default unless there is a prompt and the user selects Yes.

This change will allow the user to disable this feature when
the message prompt is disabled.

Perhaps a future enhancement is to add another checkbox in
Tools |Options | Project and Solutions | General | "Always pply symbol rename"
so this will allow the user to apply symbol rename without the prompt.