Improve first up-to-date failure log message
The up-to-date check must return false for the first time it is executed against a project. This ensures that changes made to the project since it was last loaded (such as items being deleted) are not missed.

Previously these failures were accompanied by a message resembling:

 > Input 'MyFile' (2/10/2020 12:34:56) has been modified since the last up-to-date check (01/01/0001 00:00:00), not up to date.

The "zero" date is DateTime.MinValue.

With the change in this commit, this scenario now logs:

> The up-to-date check has not yet run for this project. Not up-to-date.