Abstract away creation of IPropertyPageQueryCache
Our Project Query API implementation frequently needs to access project configurations and rules. To speed access to these we cache them inside `PropertyPageQueryCache`. The operation of this type is rather mechanical, and as such is not very interesting from the point of view of testing the Project Query API implementation. At the same time it depends on multiple project services, and setting up mocks is cumbersome.

To simplify this, we previously abstracted away various uses of `PropertyPageQueryCache` with the `IPropertyPageQueryCache` interface. However, this still left a number of scenarios that were difficult to test as they depended on instantiating and using a `PropertyPageQueryCache`, and so they still had the same difficulties with adding mocks.

Here we go a step further and abstract away the creation of `IPropertyPageQueryCache` instances with the help of a new interface, `IPropertyPageQueryCacheProvider`. The "product" version of this is very simple: it just creates and returns a `PropertyPageQueryCache`. A future change will utilize this to make it easier to create mock `IPropertyPageQueryCache`s.