Extract core logic
Extract out the core logic of setting a property within the matching configurations of a project from the traversal of the Project Query API types. `ProjectSetUIPropertyValueActionCore` has no dependencies on Project Query API types and is responsible for finding and updating the actual properties. On the other hand, `ProjectSetUIPropertyValueActionBase<T>` and its derived types are responsible for extracting the necessary data from the Project Query API types, and delegating to `ProjectSetUIPropertyValueActionCore`.

There are a couple of benefits to this: the types each have fewer responsibilities, and it is easier to write tests for the core logic as we don't need to wrap up the test data in the Project Query APIs types.