Simplify tests
Simplify our test infrastructure around the various `*PropertiesAvailableStatus` types. Currently, when we create a mock `*PropertiesAvailableStatus` we mark all properties as included by default. This is the opposite of how it should be for most tests. There are many cases where we don't care about the properties at all, and don't need to include them. There are also cases where we need to be able to include just one or two properties (either to test them in isolation or just to emphasize that they are the relevant properties for the purpose of the test).

Here we update the relevant factory methods to account for this. Each now takes a boolean, `includeAllProperties`, as the first parameter and defaults it to `false`. The other parameters, one per property, are now nullable booleans. If the unit tests indicates that we should (or should not) include a particular property by passing the appropriate parameter we will use that; otherwise we'll fallback to the value of the `includeAllProperties` parameter. This allows us to eliminate a number of factory methods that are no longer needed, and to simplify a number of tests. Finally, if new properties are added in the future we won't try to include them by default, which would risk breaking existing tests that don't provide values for those properties.