Update Microsoft.VSSDK.BuildTools
Update Microsoft.VSSDK.BuildTools to a version that properly runs
32-bit-only tasks in a 32-bit environment. In the past this hasn't
mattered, as they were run inside the VS process and VS was always
32-bit. In Codespaces the VS server process is 64-bit, and these tasks
would fail to load. In the new SDK the tasks/targets have been updated
to properly indicate that they need to run in a 32-bit MSBuild process.

This should allow us to properly build our repo within a Codespace.